+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Release notes
 =============
 
+Version 0.8.68
+--------------
+
+IMPROVEMENT
+
+* Support 32-bit ARM/Intel processors
+
+BUG FIX
+
+* Fix #312 integer overflow in RESP parser
+* Prohibit use of WAL with single volume
+
 Version 0.8.67
 --------------